Move and refactor resolve_config
Why:

The functionalities of resolve_config while strongly related to `cli`
are mostly related to parsing of configuration. All builders are
currently implemented in `orion.core.io`, including
`orion.core.experiment_builder` which will make heavy use of
`resolve_config`. Furthermore, by refactoring the experiment building
process, most of the use of `resolve_config` will be moved from `cli`
to `orion.core.experiment_builder`.

How:

Move file `orion.core.cli.resolve_config` to `orion.core.io.resolve_config`.
Move class `OrionArgParser` with functions `get_basic_args_group` and
`get_user_args_group` to new submodule `orion.core.cli.base`.

All `cli` modules are modified accordingly.
orion/core/cli/__init__.py
orion/core/cli/hunt.py
orion/core/cli/insert.py
orion/core/cli/init_only.py
orion/core/cli/base.py
